+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 <!-- CHANGELOG NEWEST -->
 - IMPROVE performance of insert-many to the memory RxStorage by 40%
+- IMPROVE performance of `getHeightOfRevision()`
 - ADD check to throw helpfull error message when a Date() object is passed instead of plain json data.
 <!-- ADD new changes here! -->